+++ service_container.go
@@ -14,15 +14,9 @@ type ServiceContainer interface {
 	// error for a service not to be registered to this key.
 	Get(key interface{}) (interface{}, error)
 
-	// MustGet calls Get and panics on error.
-	MustGet(service interface{}) interface{}
-
 	// Set registers a service with the given key. It is an error for
 	// a service to already be registered to this key.
 	Set(key interface{}, service interface{}) error
-
-	// MustSet calls Set and panics on error.
-	MustSet(service interface{}, value interface{})
 
 	// Inject will attempt to populate the given type with values from
 	// the service container based on the value's struct tags. An error
@@ -76,16 +70,6 @@ func (c *serviceContainer) Get(key interface{}) (interface{}, error) {
 	return nil, fmt.Errorf("no service registered to key %s", prettyKey(key))
 }
 
-// MustGet calls Get and panics on error.
-func (c *serviceContainer) MustGet(service interface{}) interface{} {
-	value, err := c.Get(service)
-	if err != nil {
-		panic(err.Error())
-	}
-
-	return value
-}
-
 // Set registers a service with the given key. It is an error for
 // a service to already be registered to this key.
 func (c *serviceContainer) Set(key interface{}, service interface{}) error {
@@ -106,13 +90,6 @@ func (c *serviceContainer) Set(key interface{}, service interface{}) error {
 	c.services[key] = service
 
 	return nil
-}
-
-// MustSet calls Set and panics on error.
-func (c *serviceContainer) MustSet(service interface{}, value interface{}) {
-	if err := c.Set(service, value); err != nil {
-		panic(err.Error())
-	}
 }
 
 // Inject will attempt to populate the given type with values from