+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Change type of service keys from `string` to `interface{}`. [#4](https://github.com/go-nacelle/service/pull/4)
 
+### Removed
+
+- Removed MustGet and MustSet methods from the Container interface. [#3](https://github.com/go-nacelle/service/pull/3)
+
 ## [v1.0.2] - 2020-09-30
 
 ### Removed